+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/forwarders/SimpleForwarder.kt
@@ -1,7 +1,7 @@
 package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.forwarders
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.PostMessage
-import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.executeSync
+import com.github.insanusmokrassar.AutoPostTelegramBot.utils.extensions.executeBlocking
 import com.pengrad.telegrambot.TelegramBot
 import com.pengrad.telegrambot.model.Message
 import com.pengrad.telegrambot.request.ForwardMessage
@@ -15,7 +15,7 @@ class SimpleForwarder : Forwarder {
         return true
     }
 
-    override fun forward(bot: TelegramBot, targetChatId: Long, vararg messages: PostMessage): Map<PostMessage, Message> {
+    override suspend fun forward(bot: TelegramBot, targetChatId: Long, vararg messages: PostMessage): Map<PostMessage, Message> {
         return messages.mapNotNull {
             postMessage ->
             val message = postMessage.message ?: return@mapNotNull null
@@ -26,7 +26,7 @@ class SimpleForwarder : Forwarder {
             )
         }.map {
             (original, request) ->
-            bot.executeSync(request).let {
+            bot.executeBlocking(request).let {
                 response ->
                 response.message() ?.let {
                     original to it